Add advertisement styles, removeable home blocks
This commit both adds some styling to advertisements on the home page,
as well as changes the home page plugin to allow for modules that do not
appear, helpful for long lists of articles that need to be of variable
length. Now, we can add many items to such a list by including them in
the original home page layout; if there needs to be fewer items, we just
set them to disappear in the backend.