Change wpp_get_mostpopular range to `last30days`
The previous value, `weekly`, was invalid and was causing things like Gideon's super old article to always stay in the list, for some reason.

This value is supposedly valid? It's listed in the [plugin documentation](https://bowdoinorient.com/wp-admin/options-general.php?page=wordpress-popular-posts&tab=params) as a valid value so here's hoping. Would appreciate someone loading this up and testing it before deploying.